feat: add default timeout for time comparisons (#200)
Add a default tolerance for time comparisons that can be set via customizations, e.g.
```csharp
Customize.aweXpect.Settings().DefaultTimeComparisonTolerance.Set(15.Milliseconds());
```
This should set a default tolerance that is automatically applied to all tests (and also displayed in the test expectation). When the disposable gets disposed, the default tolerance is reset to zero.

**Motivation:**
In Windows the `DateTime` resolution is [about 10 to 15 milliseconds](https://stackoverflow.com/questions/3140826/windows-system-time-with-millisecond-precision), so comparing them as exact values results in brittle tests which could be improved by applying a corresponding default tolerance, when the tests you are doing don't require exact values.