fix: finalize release executed only after aweXpect push (#230)
With #223 the finalize release action was triggered `always`, but this leads to many unnecessary/misleading PR comments.
We consider a release only finished, after the "aweXpect" main package is released and that's the new release version, therefore the finalize release action should only be triggered on a `v{x}.{y}.{z}` tag after a successful push of the "aweXpect" package.