feat: support synchronous expectations via `.Verify()` (#245)
And add initial analyzer and code fix provider (see #175) that also supports the synchronous case.

Instead of awaiting an expectation you now have the option to add a `.Verify()` at the end. This is intended to use in cases where you cannot use `async` (e.g. when dealing with a `ref struct`).

Also add an analyzer that highlights an error, when you neither await the expectation nor call `.Verify()` at the end. The corresponding code fix provider will prepend async (as this is the normal behavior).