feat: add overwriting the global timeout locally (#352)
Add `.WithTimeout(TimeSpan)` methods to the `ExpectationResult` that allow overwriting the globally defined timespan on an individual basis.

```csharp
IAsyncEnumerable<int> myEnumerable = // ...
await Expect.That(myEnumerable).All().AreEqualTo(1)
    .WithTimeout(TimeSpan.FromSeconds(10));
```

This is especially useful in combination with the `ExecutesWithin` method, as it will allow cancelling the (potentially long-running) task when the expected timeout is exceeded.